week4/scripts/requirements.txt
matplotlib==3.10.7
pandas==2.3.3
PyYAML==6.0.2
PyYAML==6.0.3
scikit_learn==1.7.2
torch==2.8.0+cu129
torchvision==0.23.0+cu129
